Make `--watch` deal with configuration changes.
This change means that you can run `bundle / build --watch` and have it
respect config changes.

If in tab one you have:

  jspm bundle app --watch

And then in another you run:

  jspm install jquery

When you change your app to `import $ from jQuery` it will pick up the
change and include jQuery in the generated bundle. Previously at this
point it would error and complain that it couldn't find jQuery, because
the config hadn't been reloaded.