inspect: multiple versions detail
By running the command `jspm inspect`, we can check if there are
multiple versions of the same dependency in the project.

This commit adds a new flag --multiple that allows to have more
details about these multiple versions, notably from which parent
dependency they come from.
This is super useful when trying to understand why there are
multiple versions of the same dependency in the project.

Example output

     Installed Versions

            npm:d3 3.5.17 npm:d3-heatmap@1.2.1
            npm:d3 4.1.0
       npm:isarray 0.0.1 npm:readable-stream@1.1.14
       npm:isarray 1.0.0 npm:buffer@3.6.0

We can see that there are 2 versions of d3:
* 3.5.17, which comes from the parent dependency d3-heatmap
* 4.1.0, which is a direct dependency (as it has no parent)