refactor(config): Implement checksum for file integrity
As it stands, JSPM 0.17 uses timestamps to check for file integrity
(i.e. checking if other processes have modified the config file).
However, this makes buildup and teardown of JSPM enviroments a pain
since the jspm config cannot be modified in any way. I believe it's
fair to be able to modify the file as long as the contents remain
exactly the same. A decent way to accomplish this is with checksum.